# [ opt ] cubical: BoolSet for Set Bool and IntMap for Map Int