Small optimization (attempt) in code for Cubical Agda: 
- dedicated type `BoolSet` for `Set Bool`
- use `IntMap` instead of `Map Int`