@nad 2022-04-26
> Small optimization (attempt) in code for Cubical Agda:
> 
> * dedicated type `BoolSet` for `Set Bool`

Do you plan to benchmark this change?