@andreasabel 2022-04-26
> Do you plan to benchmark this change?

Ah, indeed, at least at the library level this is benchmarkable...