@andreasabel 2022-04-27
Initial benchmarking confirms the expected: the enum type `BoolSet` is faster than `Set Bool`. 
This is for GHC 9.2.2, but GHC 8.0.2 is similar:
```
All
  insert
    Set Bool: OK (1.77s)
      834  ns ±  56 ns
    BoolSet:  OK (1.15s)
      137  ns ±  13 ns
  subset fromList
    Set Bool: OK (3.08s)
      45.4 ns ± 1.9 ns
    BoolSet:  OK (1.48s)
      21.6 ns ± 1.8 ns
  subset fromAscList
    Set Bool: OK (6.17s)
      45.5 ns ± 1.2 ns
    BoolSet:  OK (1.58s)
      23.3 ns ± 1.8 ns
  subset fromDistinctAscList
    Set Bool: OK (2.15s)
      7.92 ns ± 430 ps
    BoolSet:  OK (1.86s)
      6.88 ns ± 664 ps
```
These are the benchmarks:
```haskell
main :: IO ()
main = defaultMain
  [ bgroup "insert"
    [ bench "Set Bool" $ nf (bench_insert_False     Set.empty     Set.insert     Set.lookupMin) 100
    , bench "BoolSet"  $ nf (bench_insert_False BoolSet.empty BoolSet.insert BoolSet.lookupMin) 100
    ]
  , bgroup "subset fromList"
    [ bench "Set Bool" $ nf (bench_subset_fromList     Set.isSubsetOf     Set.fromList) True
    , bench "BoolSet"  $ nf (bench_subset_fromList BoolSet.isSubsetOf BoolSet.fromList) True
    ]
  , bgroup "subset fromAscList"
    [ bench "Set Bool" $ nf (bench_subset_fromAscList     Set.isSubsetOf     Set.fromAscList) False
    , bench "BoolSet"  $ nf (bench_subset_fromAscList BoolSet.isSubsetOf BoolSet.fromAscList) False
    ]
  , bgroup "subset fromDistinctAscList"
    [ bench "Set Bool" $ nf (bench_subset_fromDistinctAscList     Set.isSubsetOf     Set.fromDistinctAscList) ()
    , bench "BoolSet"  $ nf (bench_subset_fromDistinctAscList BoolSet.isSubsetOf BoolSet.fromDistinctAscList) ()
    ]
  ]

bench_insert_False empty insert lookupMin n =
  lookupMin (foldr insert empty $ replicate n False)
  ==
  Just False

bench_subset_fromList isSubsetOf fromList b =
  fromList [False]
  `isSubsetOf`
  fromList [False,b]

bench_subset_fromAscList isSubsetOf fromAscList b =
  fromAscList [False]
  `isSubsetOf`
  fromAscList [False,b]

bench_subset_fromDistinctAscList isSubsetOf fromDistinctAscList () =
  fromDistinctAscList [False]
  `isSubsetOf`
  fromDistinctAscList [False,True]
```
My curiosity is satisfied with that. 
@nad: If you want to try and refute it, I am happy to provide the full benchmarking sources.