@nad 2022-04-27
Does the new module make Agda so much faster that it outweighs the extra (presumably small) maintenance cost?