@andreasabel 2022-04-27
> Does the new module make Agda so much faster that it outweighs the extra (presumably small) maintenance cost?

I don't know how to quantify that.  `Set Bool` just hurt my eyes, so I was compelled to replace it.
Since `BoolSet` is a drop-in replacement for `Set Bool`, I think that the extra cost is mostly recompilation time, but this shouldn't be so much as I expect the `BoolSet` module to be stable.  (I have already written `SmallSet` with a similar motivation.)
I could of course publish `BoolSet` so that we just have it as dependency, but then it would have to be more professional implementing the _whole_ `Data.Set` interface...